movieList/requirements.txt
certifi==2019.6.16
chardet==3.0.4
cssselect==1.0.3
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-filters==0.2.1
djangorestframework==3.9.4
idna==2.8
lxml==4.3.4
mysqlclient==1.4.2.post1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
